Core/Spell: Implement SMSG_SEND_SPELL_HISTORY & CAST_FLAG_EX_TOY_COOLDOWN. Fix ToyBox cooldowns.

+++ src/server/game/Entities/Player/Player.h
@@ -230,7 +230,7 @@ typedef std::list<uint64> WhisperListContainer;
 struct SpellCooldown
 {
     uint64 end;
-    uint16 itemid;
+    uint32 itemid;
 };
 
 typedef std::map<uint32, SpellCooldown> SpellCooldowns;
